+++ src/drivers/bublbobl.c
@@ -100,7 +100,7 @@ CPU 3
 
 /* vidhrdw/bublbobl.c */
 extern unsigned char *bublbobl_objectram;
-extern int bublbobl_objectram_size;
+extern size_t bublbobl_objectram_size;
 void bublbobl_vh_convert_color_prom(unsigned char *palette, unsigned short *colortable,const unsigned char *color_prom);
 void bublbobl_vh_screenrefresh(struct osd_bitmap *bitmap,int full_refresh);
