+++ src/drivers/ccastles.c
@@ -97,7 +97,7 @@ WRITE_HANDLER( ccastles_flipscreen_w );
 
 
 static unsigned char *nvram;
-static int nvram_size;
+static size_t nvram_size;
 
 static void nvram_handler(void *file,int read_or_write)
 {
