+++ src/drivers/cloud9.c
@@ -66,7 +66,7 @@ extern unsigned char *cloud9_color_bank;
 
 
 static unsigned char *nvram;
-static int nvram_size;
+static size_t nvram_size;
 
 static void nvram_handler(void *file,int read_or_write)
 {
