+++ src/drivers/gladiatr.c
@@ -262,7 +262,7 @@ static READ_HANDLER( glad_cpu_sound_command_r )
 
 
 static unsigned char *nvram;
-static int nvram_size;
+static size_t nvram_size;
 
 static void nvram_handler(void *file,int read_or_write)
 {
