+++ src/drivers/kopunch.c
@@ -2,7 +2,7 @@
 #include "vidhrdw/generic.h"
 
 unsigned char *bsvideoram;
-int bsvideoram_size;
+size_t bsvideoram_size;
 
 void kopunch_vh_convert_color_prom(unsigned char *palette, unsigned short *colortable,const unsigned char *color_prom)
 {
