+++ src/drivers/travrusa.c
@@ -17,7 +17,7 @@ Loosely based on the our previous 10 Yard Fight driver.
 #include "sndhrdw/irem.h"
 
 extern unsigned char *spriteram;
-extern int spriteram_size;
+extern size_t spriteram_size;
 
 extern unsigned char *travrusa_videoram;
