+++ src/vidhrdw/generic.h
@@ -2,14 +2,14 @@
 
 
 extern unsigned char *videoram;
-extern int videoram_size;
+extern size_t videoram_size;
 extern unsigned char *colorram;
 extern unsigned char *spriteram;
-extern int spriteram_size;
+extern size_t spriteram_size;
 extern unsigned char *spriteram_2;
-extern int spriteram_2_size;
+extern size_t spriteram_2_size;
 extern unsigned char *spriteram_3;
-extern int spriteram_3_size;
+extern size_t spriteram_3_size;
 extern unsigned char *buffered_spriteram;
 extern unsigned char *buffered_spriteram_2;
 extern unsigned char *flip_screen;
